+++ core/helper_test.go
@@ -3,9 +3,11 @@ package core
 import (
 	"crypto/rand"
 	"errors"
+	"fmt"
 	"io"
 	"math/big"
 	"net/http"
+	"reflect"
 	"strings"
 	"testing"
 
@@ -64,16 +66,16 @@ func TestToQueryMap(t *testing.T) {
 	}{
 		{
 			Input: struct {
-				Param1 string `query:"param_1"`
-				Param2 int    `query:"param_2"`
-				Param3 bool   `query:"param_3"`
+				Param1 string `url:"param_1"`
+				Param2 int    `url:"param_2"`
+				Param3 bool   `url:"param_3"`
 			}{
 				Param1: "value1",
 				Param2: 5,
 				Param3: true,
 			},
 			ShouldError: false,
-			Expected:    map[string]string{"param_1": "value1", "param_2": "5", "param_3": "1"},
+			Expected:    map[string]string{"param_1": "value1", "param_2": "5", "param_3": "true"},
 			CheckErr:    nil,
 		},
 		{
@@ -81,6 +83,14 @@ func TestToQueryMap(t *testing.T) {
 			ShouldError: true,
 			Expected:    nil,
 			CheckErr:    func(err error) bool { return assert.Equal(t, ErrNilSource, err) },
+		},
+		{
+			Input:       "test",
+			ShouldError: true,
+			Expected:    nil,
+			CheckErr: func(err error) bool {
+				return assert.Equal(t, fmt.Errorf("query: Values() expects struct input. Got %v", reflect.String), err)
+			},
 		},
 	}