+++ table-api/table_item_request_builder_get_query_parameters.go
@@ -12,7 +12,7 @@ type TableItemRequestBuilderGetQueryParameters struct {
 	//- Encrypted text: The database value is encrypted, while the displayed value is unencrypted based on the user's encryption context.
 	//
 	//- Reference fields: The database value is sys_id, but the display value is a display field of the referenced record.
-	DisplayValue DisplayValue `query:"sysparm_display_value"`
+	DisplayValue DisplayValue `url:"sysparm_display_value"`
 	//ExcludeReferenceLink flag that indicates whether to exclude Table API links for reference fields.
 	//
 	//Valid values:
@@ -20,9 +20,9 @@ type TableItemRequestBuilderGetQueryParameters struct {
 	//- true: Exclude Table API links for reference fields.
 	//
 	//- false: Include Table API links for reference fields.
-	ExcludeReferenceLink bool `query:"sysparm_exclude_reference_link"`
+	ExcludeReferenceLink bool `url:"sysparm_exclude_reference_link"`
 	// Fields list of fields to return in the response.
-	Fields []string `query:"sysparm_fields"`
+	Fields []string `url:"sysparm_fields"`
 	//Flag that indicates whether to restrict the record search to only the domains for which the logged in user is configured.
 	//
 	//Valid values:
@@ -30,7 +30,7 @@ type TableItemRequestBuilderGetQueryParameters struct {
 	//- false: Exclude the record if it is in a domain that the currently logged in user is not configured to access.
 	//
 	//- true: Include the record even if it is in a domain that the currently logged in user is not configured to access.
-	QueryNoDomain bool `query:"sysparm_query_no_domain"`
+	QueryNoDomain bool `url:"sysparm_query_no_domain"`
 	// View	UI view for which to render the data. Determines the fields returned in the response.
 	//
 	//Valid values:
@@ -39,5 +39,5 @@ type TableItemRequestBuilderGetQueryParameters struct {
 	//- mobile
 	//- both
 	//If you also specify the sysparm_fields parameter, it takes precedent.
-	View View `query:"sysparm_view"`
+	View View `url:"sysparm_view"`
 }